fix(glossary): extract history actor fallback for isolated tests
Move actor display-name resolution out of the history page query so unit tests do not load database env, and keep the regenerated invalid-cursor message id.
